Remove "buffer" module
It wasn't pulling its weight.  Move its contents into lib.rs instead.
